--- src/utils/createday.js
@@ -1,22 +1,0 @@
-export const formatDay = (iso) => {
-  if (!iso) return "";
-
-  const date = new Date(iso);
-  const now = new Date();
-
-  const isToday =
-  date.getFullYear() === now.getFullYear() &&
-  date.getMonth() === now.getMonth() &&
-  date.getDate() === now.getDate();
-
-  if (isToday) {
-    const hours = String(date.getHours()).padStart(2, "0");
-    const minutes = String(date.getMinutes()).padStart(2, "0");
-    return `${hours}:${minutes}`;
-  } else {
-    const year = date.getFullYear();
-    const month = String(date.getMonth() + 1).padStart(2, "0");
-    const day = String(date.getDate()).padStart(2, "0");
-    return `${year}-${month}-${day}`;
-  }
-};
