+++ src/utils/index.js
@@ -1,6 +1,4 @@
-export * from "./format";
-export * from "./path";
-export * from "./date";
-export * from "./createday";
+export * from "./buildPath";
 export * from "./parseContent";
-export * from "./apiHandler";
+export * from "./handleApiCall";
+export * from "./format";
